+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Added parsing of Null,NULL,~
+- Added parsing of False,FALSE
+- Added parsing of True,TRUE
+
+### Changed
+
+- Parses integers returns value as int type
+- Parses floats returns value as float type
 
 ## [1.1.3] - 2020-07-26